Try to get tests passing for CC tbl stuff.
Some of the changes I'm not so sure about, like why we need to
invalidate the gccct table when freeing a class/module. I actually think
we need to invalidate the gccct table for the given class/module when we
do a move during compaction, because the gccct table is keyed on the
pointer itself so it could get the wrong CC and cc->klass could == klass
and still match (I think).

Overall I'm not so happy with these changes but they at least allow us
to move forward right now with actually implementing RCU instead of locks and
benchmarking the results.